% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGeneSignatures.R
\name{rankmatrix}
\alias{rankmatrix}
\title{For the genes being present in genes.ranged per celltype we rank from 1 (best rank) to length 
of the current genes.ranged entry, and NA for genes not in that respective entry but in u:}
\usage{
rankmatrix(genes.ranged)
}
\description{
For the genes being present in genes.ranged per celltype we rank from 1 (best rank) to length 
of the current genes.ranged entry, and NA for genes not in that respective entry but in u:
}
